+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 Due to package verification, the latest version below is the unpublished version and the date is meaningless.
 however, it has to be formatted properly to pass verification tests.
 
+## [0.9.4-preview] - 2019-8-20
+
+### Fixed
+### Actions
+### Changed
+### Added
+
 ## [0.9.3-preview] - 2019-8-15
 
 ### Fixed